backend/python/requirements.txt
# ============================================================
# requirements.txt — Python dependencies cho Freshness Server
# Yêu cầu: Python >= 3.9, < 3.13
# Cài đặt: pip install -r requirements.txt
# ============================================================

# ── Web framework ────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── Roboflow Inference SDK (HTTP client gọi Roboflow API) ───────
# inference-sdk: gọi API Roboflow để chạy model từ xa (nhẹ, ít dependency)
# Cung cấp: from inference_sdk import InferenceHTTPClient
inference-sdk>=1.0.0

# ── Xử lý ảnh ───────────────────────────────────────────────
Pillow==11.0.0
numpy>=2.0.0

# ── Hỗ trợ multipart form / file upload ─────────────────────
python-multipart==0.0.12

# ── HTTP client (để inference SDK gọi Roboflow server) ──────
httpx==0.27.2
requests>=2.28.0

# ── (Tuỳ chọn) GPU acceleration với ONNX runtime ────────────
# Bỏ comment dòng dưới nếu có GPU NVIDIA:
# onnxruntime-gpu==1.19.2
# Hoặc chỉ dùng CPU:
onnxruntime==1.19.2
